added new fuzzy logic controller and added log4j2 config and code refactoring

+++ src/main/java/pl/uam/movieSelector/spring/service/FuzzyLogicService.java
@@ -11,6 +11,6 @@ import java.util.ArrayList;
 
 public interface FuzzyLogicService {
 
-    ArrayList<MovieData> predictUserAnswers(final ArrayList<UserQuestionData> userQuestions, final int nTopMovies);
+    ArrayList<MovieData> predictUserAnswers(final ArrayList<UserQuestionData> userQuestions, final long nTopMovies);
 
 }
